requirements.txt
black==23.3.0
click==8.1.3
pydantic==2.0
PyYAML==6.0
psycopg2-binary==2.9.6